Use script directory in rtp when running from git
Otherwise the script can only be run from within the git directory.  With this
change it can also be run from anywhere else by giving an absolute or relative
path.

It is assumed that the project directory is not in $PATH so there will always
be at least one slash in $0 when calling nvimpager.in.