Use bash's $PPID to find the calling process
This simplifies the logic to find the parent process of the pager which
is checked.  But it also disables the recursive upward search in the
process tree.  If nvimpager is called by man, git, pydoc and friends
directly this is not a problem.  But a wrapper around nvimpager has to
take care that the command name of nvimpager's parent process can be
identified (either by exec-ing nvimpager or by changing their own
command name).