Fix reading from stdin for nvim>=0.3.0
Neovim version 0.3.0 accepts text from stdin even if - was not given as
argument.  This means that it reacts differently to the open file
descriptor or pipeline on stdin when we read from stdin.  By connecting
neovims stdin to the terminal explicitly this can be avoided.

When https://github.com/neovim/neovim/pull/8246 is merged we can try to
rewrite the --headless switch and pipe text into nvim directly.