Use new lua function and fix tests
The old test fixtures were created for the vimscript version of
group2ansi().  The old function did reset all terminal attributes
individually with the escape sequence "22;24;25;27;28;39;49", the
"reset-all" sequence "0" is used in the lua version as I have not found
any problems with it so far.

The makefile now also installs the lua file.