Use the new lua function for highlighting
The resetting of the terminal attributes after the last line of output
is missing from this version.

The test cases are adapted to the new escape sequences.  The test cases
for concealing are marked as pending until that is implemented in lua.

The old vimscript version is kept for reference.