Correctly check return status in tests
The io.close function in lua 5.1 only returns true.  It does not return
the return status of the executed command.  In order to get that it has
to be printed by the command (shell) itself and the parsed in lua.

Currently we want to support lua 5.1 because that is what neovim uses.