Add test for parent process detection
These tests exercise the lua function directly from nvim.  This should
help to prevent false negatives.  The wrapper script was updated to
communicate the PPID environment variable to a child process.  This is
needed to call nvim directly as the variable is normally set up by the
nvimpager bash script.