Move modeline to end of test file
This commit does not make any relevant changes to the project itself.
It solely exists to provide git diff and git log output that can be used
to test the pager with modelines from diffed files.

The modeline is moved to the end in order to make it appear on the last
line of the generated diff and hence to be picked up by neovim even
with a git diff header in the output.