Detect parent process during --cmd
The parent process needs to be detected early to prevent modelines in
git diffs from being read.  This would happen before the -c commands are
run so it has to be done in the --cmd commands already.

The result of the detection is needed later on so it is saved in a
(module) variable.