Only trap EXIT for cleanup
When we trap any of the other signals we "handle" them and the script
continues to run afterwards because the handler did not exit.  When we
only handle EXIT all other signals trigger their normal effects and only
these that force the shell to exit then trigger the trap.

This is related to #17 where aborting `./nvimpager` (without arguments)
by pressing CTRL-C resulted in an error message from the `cat $files |
wc` part because the files where not there.  The underlying problem was
that the script did not exit on CTRL-C because the trap did handle and
ignore the signal.