Add option to use lowdown instead of pandoc for man page generation.
We use a make variable to select the markdown processor. Pandoc is quite
big (needs Haskell and many Haskell packages to compile) but present in
many repositories, lowdown is very small and aims to be very portable
(small C program) but it is not present in many repositories.

This patch adopts metadata in the style supported by lowdown (at least
for now, YAML blocks might yet be supported), with the -m command line
options that are provided since version 0.8.1 of lowdown.  It also makes
authors section in the man page explicit, for portability between pandoc
and lowdown.

Also create a DATE variable, which can be overriden when building this
in environments without git available, or when building from a tarball.

Co-authored-by: Lucas Hoffmann <lucc@posteo.de>