Change order of nvimpager's and user supplied options
Previously the stage1() was run after all user supplied --cmd arguments
and the stage2() was run after all user supplied -c arguments.  This
made it very complicated to correctly test pager mode in the new test
framework.

This is a breaking change for all users who depend on stage1 or stage2
to overwrite their --cmd or -c settings.  I would expect that the
predominant expectation is to overwrite settings from nvimpager in user
supplied options and not the other way around.